requirements.txt
pdfplumber
pytesseract
Pillow
langdetect